+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Change Log
 - Can now deal with exceptions that can't be serialized: they're raised as generic PyroError instead with appropriate message
 - added new config item FLAME_ENABLED, to enable/disable the use of Pyro Flame on the server. Default is false (disabled).
 - Moved futures from core to new futures module. Code using Pyro4.Future will still work.
+- Added python version info to configuration dump
 
 
 **Pyro 4.16**